+++ docs/changelog.rst
@@ -19,6 +19,7 @@ Changelog
 * Fixed a crash when reloading unloaded or unsupported resources.
 * Added support to shaders and materials hot-reloading.
 * Data Compiler: #defines are now correctly exposed in vs_input_output and fs_input_output.
+* Lua API: fixed ``Math.ray_obb_intersection()`` with scaled OBBs.
 
 0.54.0 --- 13 Jan 2025
 ----------------------